https://github.com/singnet/snet-daemon/issues/286 Moved the price to an independent package Goal is to support method level pricing and also fixed pricing in the current state For future we will have more pricing types
