+++ escrow/income_test.go
@@ -2,7 +2,7 @@ package escrow
 
 import (
 	"fmt"
-	price2 "github.com/singnet/snet-daemon/blockchain/price"
+	price2 "github.com/singnet/snet-daemon/price"
 	"github.com/singnet/snet-daemon/handler"
 	"math/big"
 	"testing"
