+++ price/fixed_method_pricing.go
