+++ price/fixed_method_pricing_test.go
@@ -14,14 +14,14 @@ import (
 
 
 func TestFixedMethodPrice_initPricingData(t *testing.T) {
-	metadata,_ := blockchain.ReadServiceMetaDataFromLocalFile("../../service_metadata_method_pricing.json")
+	metadata,_ := blockchain.ReadServiceMetaDataFromLocalFile("../service_metadata_method_pricing.json")
 	grpcCtx := &handler.GrpcStreamContext{Info:&grpc.StreamServerInfo{FullMethod:"/example_service.Calculator/add"}}
 	pricing,_ := InitPricing(metadata)
-	price,err := pricing.pricingTypes[0].GetPrice(grpcCtx)
+	price,err := pricing.GetPrice(grpcCtx)
 	assert.Equal(t,price,big.NewInt(2))
 	assert.Nil(t,err)
 	grpcCtx.Info.FullMethod= "NonDefinedMethod"
-	price,err = pricing.pricingTypes[0].GetPrice(grpcCtx)
+	price,err = pricing.GetPrice(grpcCtx)
 	assert.Nil(t,price)
 	assert.Equal(t,err.Error(),"price is not defined for the Method NonDefinedMethod")
